Split to separate modules, unload morseword and run garbage collect before loading pwmenu with gc free mem printouts
